+++ src/utils/get-test-and-fixture-info.ts
@@ -1,13 +1,13 @@
 import TestRun from '../test-run';
 
-export function getTestRunTestInfo (testRun: TestRun): object {
+export function getTestInfo (testRun: TestRun): object {
     return {
         name: testRun.test.name,
         meta: { ...testRun.test.meta },
     };
 }
 
-export function getTestRunFixtureInfo (testRun: TestRun): object {
+export function getFixtureInfo (testRun: TestRun): object {
     return {
         name: testRun.test.fixture?.name,
         path: testRun.test.fixture?.path,
