+++ ts-defs-src/test-api/test-info.d.ts
@@ -1,0 +1,15 @@
+// Test Info
+//----------------------------------------------------------------------------------------------------------------------
+
+type Metadata = Record<string, string>;
+
+interface TestInfo {
+    name: string;
+    meta: Metadata;
+}
+
+interface FixtureInfo {
+    name: string;
+    meta: Metadata;
+    path: string;
+}
